backend/requirements.txt
fastapi
uvicorn
requests
pydantic
python-dotenv
openai
httpx
pdfplumber
google-generativeai
